plumbing: format/index, add ModTime field to Index struct
Store the modification time of the index file in the Index structure.
This timestamp is captured when the index is loaded and will be used
to properly handle the racy-git condition in the worktree status
optimization.

The ModTime is obtained by calling Stat() on the open file handle,
ensuring it corresponds exactly to the index content that was read.

Reference: https://git-scm.com/docs/racy-git